spark-app/requirements.txt
kafka-python
pyspark
elasticsearch
astral
